REFACT(PLOT): MOVE PLOT in own module
