+++ test/test_plot.py
@@ -1,12 +1,12 @@
 # Copyright 2016, Yahoo Inc.
 # Licensed under the terms of the Apache License, Version 2.0. See the LICENSE file associated with the project for terms.
 
+import sys
 from operator import add
 
 import pytest
-import sys
 
-from graphkit import compose, network, operation
+from graphkit import base, compose, network, operation, plot
 from graphkit.modifiers import optional
 
 
@@ -48,7 +48,9 @@ def solution(pipeline, inputs, outputs, request):
 def test_plotting_docstring():
     common_formats = ".png .dot .jpg .jpeg .pdf .svg".split()
     for ext in common_formats:
-        assert ext in network.plot_graph.__doc__
+        assert ext in plot.plot_graph.__doc__
+        assert ext in base.NetworkOperation.plot.__doc__
+        assert ext in network.Network.plot.__doc__
 
 
 def test_plot_formats(pipeline, input_names, outputs, solution, tmp_path):
